porterias y goles, ademas de botes y botin

+++ principal.js
@@ -22,6 +22,7 @@ var ball_velocity;
 var mando; 
 var salto;
 var gol;
+var sePuedeMarcar = true;
 
 
 //
@@ -33,12 +34,14 @@ var estado_princ = {
     
     preload: function(){
         capgross.load.image('fondo', 'img/campo.png');
-        capgross.load.image('pelota', 'img/Pelota2.png');
+        capgross.load.image('pelota', 'img/Pelota3.png');
         capgross.load.image('jugador1', 'img/EricR.png');
         capgross.load.image('jugador2', 'img/JordiR.png');
         capgross.load.image('Miniatura1', 'img/EricR.png');
         capgross.load.image('Miniatura2', 'img/JordiR.png');
         capgross.load.image('Marcador', 'img/Marcador.png');
+        capgross.load.image('Botin', 'img/botinBasic.png');
+        capgross.load.image('Largero', 'img/Largero.png');
     },
     
 
@@ -46,7 +49,16 @@ var estado_princ = {
         capgross.add.tileSprite(0, 0, 1074, 724, 'fondo');
         personaje = capgross.add.sprite(capgross.width, capgross.height, 'jugador1');
         personaje2 = capgross.add.sprite(capgross.width/2, capgross.height/2, 'jugador2');
+        botin = capgross.add.sprite(capgross.width-100, capgross.height-50,'Botin');
+        largero = capgross.add.sprite(capgross.width, capgross.height/2 +150,'Largero');
+        largero2 = capgross.add.sprite(0, capgross.height/2 +150,'Largero');
         
+        
+        largero.scale.setTo(-0.35,0.25);
+        largero2.scale.setTo(0.25);
+        
+        botin.angle = 25;
+
         ball = create_ball(capgross.world.centerX,capgross.world.centerY);
 
 
@@ -70,27 +82,34 @@ var estado_princ = {
         ball_velocity = 400; 
 
 
-
+        
         capgross.physics.arcade.enable(personaje, ball);
         capgross.physics.arcade.enable(personaje2, ball);
-
+        capgross.physics.arcade.enable(botin, ball);
+        capgross.physics.arcade.enable(botin, personaje2);
 
 
         cursor = capgross.input.keyboard.createCursorKeys();
         derecha = capgross.input.keyboard.addKey(Phaser.Keyboard.D);
         izquierda = capgross.input.keyboard.addKey(Phaser.Keyboard.A);
         arriba = capgross.input.keyboard.addKey(Phaser.Keyboard.W);
-
+        space = capgross.input.keyboard.addKey(Phaser.Keyboard.SPACEBAR);
+        
         capgross.physics.startSystem(Phaser.Physics.ARCADE);
         
         
         personaje.body.gravity.y = 1900;
         personaje2.body.gravity.y = 1900;   
-        
+        botin.body.gravity.y = 1900;
    
 
         personaje.body.collideWorldBounds = true;
         personaje2.body.collideWorldBounds = true;
+        botin.body.collideWorldBounds = true;
+
+
+
+
 
         //miniaturas y scoreboard
         puntos1 = 0;
@@ -127,48 +146,65 @@ var estado_princ = {
     ///
     },
     update: function(){
-/* celebracion
-if (pad1.isDown(Phaser.Gamepad.XBOX360_A))
-{
-    personaje.angle += 5;
-}
-*/        
+        console.log(botin.body.velocity.y)
+        //Chute
+        if (space.isDown && botin.angle <= 135)
+        {
+            botin.angle += 5;
+        }else if (space.isUp && botin.angle >= 25)
+        {
+            botin.angle -= 5;
+        }
+        
         //Derecha
         if (cursor.right.isDown || (pad1.isDown(Phaser.Gamepad.XBOX360_DPAD_RIGHT) || pad1.axis(Phaser.Gamepad.XBOX360_STICK_LEFT_X) > 0.1)){
             personaje.position.x += 4;
+            botin.position.x += 4;
         }
         //Izqueirda
         if (cursor.left.isDown || (pad1.isDown(Phaser.Gamepad.XBOX360_DPAD_LEFT) || pad1.axis(Phaser.Gamepad.XBOX360_STICK_LEFT_X) < -0.1)){
             personaje.position.x -= 4;
-
+            botin.position.x -= 4;
         }
         //Salto
         if ((cursor.up.isDown || pad1.isDown(Phaser.Gamepad.XBOX360_X )) && personaje.body.blocked.down){
             personaje.body.velocity.y = -600;
-                      
+            botin.body.velocity.y = -600;       
         }
         if (derecha.isDown){
             personaje2.position.x += 4;
+            
         }
         if (izquierda.isDown){
             personaje2.position.x -= 4;
+            
         }
         if (arriba.isDown && personaje2.body.blocked.down){
             personaje2.body.velocity.y = -600;
+            
         }
         
+
+        /*console.log(botin.position.x +" " + ball.position.x);
+        if(botin.position.x < ball.position.x || botin.position.x+15 > ball.position.x){
+            ball.position.x -=5;
+        }*/
+
 
         capgross.physics.arcade.collide(personaje, personaje2);
         capgross.physics.arcade.collide(personaje, ball);
         capgross.physics.arcade.collide(personaje2, ball);
+        capgross.physics.arcade.collide(botin, ball);
+        capgross.physics.arcade.collide(largero, ball);
+        capgross.physics.arcade.collide(largero2, ball);
         
-        if(ball.body.blocked.left){
+        if(ball.body.blocked.left && sePuedeMarcar && ball.body.position.y > capgross.height/2 +150 ){
 
             destroySprite(miniatura1,miniatura2,txtPuntos1,txtPuntos2);
             puntos2++;
             txtPuntos2.text = puntos2;
         }
-        if(ball.body.blocked.right){
+        if(ball.body.blocked.right && sePuedeMarcar && ball.body.position.y > capgross.height/2 +150){
 
             destroySprite(miniatura1,miniatura2,txtPuntos1,txtPuntos2);            
             puntos1++;
@@ -177,6 +213,7 @@ if (pad1.isDown(Phaser.Gamepad.XBOX360_A))
 
         ball.angle += 1;
         
+        botin.body.immovable=true;
         personaje.body.immovable=true;
         personaje2.body.immovable=true;
 
@@ -194,8 +231,6 @@ function create_ball(x,y){
 }
 function launch_ball(){
     if(ball_launched){
-        ball.x = capgross.world.centerX;
-        ball.y = capgross.world.centerY;
         ball.body.velocity.setTo(0,0);
         ball_launched = false;
     }else{
@@ -209,14 +244,13 @@ capgross.state.add('principal', estado_princ);
 capgross.state.start('principal');
 
 function destroySprite (sprite,sprite1,sprite2,sprite3) {
-    
+        sePuedeMarcar = false;
         sprite.destroy();
         sprite1.destroy();
         sprite2.destroy();
         sprite3.destroy();
         gol = capgross.add.text(460, 50, "GOOOL!!!", {font:"45px Open Sans", fill:"white"});
         capgross.time.events.add(Phaser.Timer.SECOND * 2, SetMarcador, gol);
-
     }
 
 function SetMarcador() {
@@ -230,6 +264,9 @@ function SetMarcador() {
         miniatura2 = capgross.add.sprite(615, 52, 'Miniatura1');
         miniatura1.scale.setTo(-0.35,0.35);
         miniatura2.scale.setTo(0.35);
+        ball.position.x = capgross.world.centerX;
+        ball.position.y = capgross.world.centerY;
+        sePuedeMarcar = true;
     }
     
 //prueva guardar comentario
